[fs] Convert MessageID in smb2_hdr to LE
Message-id: <1421409914-15222-1-git-send-email-sprabhu@redhat.com>
Patchwork-id: 103134
O-Subject: [RHEL7.1 PATCH fs] [bz1161441]Convert MessageID in smb2_hdr to LE
Bugzilla: 1161441
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

We have encountered failures when When testing smb2 mounts on ppc64
machines when using both Samba as well as Windows 2012.

On poking around, the problem was determined to be caused by the
high endian MessageID passed in the header for smb2. On checking the
corresponding MID for smb1 is converted to LE before being sent on the
wire.

We have tested this patch successfully on a ppc64 machine.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
(cherry picked from commit 9235d09873316d602937b5d45c431fb653f3aed8)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>