[net] bridge: Program port vlan filters only if filtering is enabled in bridge
Message-id: <1421853633-6641-1-git-send-email-vyasevic@redhat.com>
Patchwork-id: 103324
O-Subject: [RHEL7 PATCH] bridge: Program port vlan filters only if filtering is enabled in bridge.
Bugzilla: 1183958
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1183958
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=8591819
Upstream: RHEL-7 only

Currently, when a vlan filter is set on a port of a bridge, the
bridge will also push this information down to the driver for
a given port.  For HW, this sets the HW vlan filters enabling
the card to receive vlans that the bridge is interested in.

After RHEL7 commit a320a5345218228875b2b1197c83c043f9eacfcf
([net] bridge: Add filtering support for default_pvid),
the above actions happen even when vlan filtering is disabled
and the user has not specified any filtering information.
This causes a regression when a port is a bond which uses
arp monitoring to track link state.  The root cause of the
regression is that the bonding driver is currently incapable
of always correctly verifying the source device of the arp if
vlans are configured above the bond.  The solution to this
problem is not trivial and has been defered till RHEL-7.2
(BZ 1166647).

This patch is here to restore the bonding behavior to
pre-a320a53452182.  It accomplishes this by setting the port
HW vlan filtering data only when vlan filtering feature
is enabled in the bridge.  This way, default bridge configuration
will not impact bond opertation.  However, if vlan filtering
is enabled on the bridge, the arp monitoring on the bond will
break again.  This functions the same as all RHEL7 kernels
prior to RHEL7 commit a320a5345218228875b2b1197c83c043f9eacfcf
([net] bridge: Add filtering support for default_pvid).

Targeting this as a RHEL7 only patch as upstream does not
have this problem.

Signed-off-by: Vladislav Yasevich <vyasevic@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>