[fs] coredump: add new %P variable in core_pattern
Message-id: <1422372933-14768-1-git-send-email-jolsa@redhat.com>
Patchwork-id: 103657
O-Subject: [PATCH RHEL7.1 BZ1186360] coredump: add new %P variable in core_pattern
Bugzilla: 1186360
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Dave Anderson <anderson@redhat.com>

From: Stephane Graber <stgraber@ubuntu.com>

Bugzilla: 1186360
https://bugzilla.redhat.com/show_bug.cgi?id=1186360

upstream
========
commit 65aafb1e7484b7434a0c1d4c593191ebe5776a2f
Author: Stephane Graber <stgraber@ubuntu.com>
Date:   Wed Sep 11 14:24:32 2013 -0700

description
===========
Add a new %P variable to be used in core_pattern.  This variable contains
the global PID (PID in the init namespace) as %p contains the PID in the
current namespace which isn't always what we want.

The main use for this is to make it easier to handle crashes that happened
within a container.  With that new variables it's possible to have the
crashes dumped into the container or forwarded to the host with the right
PID (from the host's point of view).

Cc: Jakub Filak <jfilak@redhat.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>