[misc] redhat: don't suppress Revert patches from changelog
Message-id: <1422553096-52810-1-git-send-email-jarod@redhat.com>
Patchwork-id: 103890
O-Subject: [RHEL7 PATCH] [misc] redhat: don't suppress Revert patches from changelog
Bugzilla: 1187353
RH-Acked-by: Phillip Lougher <plougher@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

We've been unwittingly suppressing patches with "Revert" in their subjects
from showing up in the changelog for some time now. Back in 2009, prior to
the release of RHEL6, patches that had names starting with ^Revert were
suppressed, and in the migration of the infrastructure to RHEL7, the ^ was
lost and the suppression made much greedier. This removes that suppression
(which was intended to filter out Fedora patches as we rebased pre-RHEL6
on whatever the current Fedora release was) and restores all the changelog
entries with Revert in them since 7.1 started up.

While we're at it, since the next build should be RC, we need to flip
RELEASED_KERNEL to 1, so that the build uses the production UEFI
secureboot keys instead of the beta ones.

Signed-off-by: Jarod Wilson <jarod@redhat.com>