[fs] nfsd: pass extra info in env vars to upcalls to allow for early grace period end
Message-id: <1423596680-6048-40-git-send-email-steved@redhat.com>
Patchwork-id: 105655
O-Subject: [RHEL7.2 PATCH 39/53] nfsd: pass extra info in env vars to upcalls to allow for early grace period end
Bugzilla: 1111712
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

In order to support lifting the grace period early, we must tell
nfsdcltrack what sort of client the "create" upcall is for. We can't
reliably tell if a v4.0 client has completed reclaiming, so we can only
lift the grace period once all the v4.1+ clients have issued a
RECLAIM_COMPLETE and if there are no v4.0 clients.

Also, in order to lift the grace period, we have to tell userland when
the grace period started so that it can tell whether a RECLAIM_COMPLETE
has been issued for each client since then.

Since this is all optional info, we pass it along in environment
variables to the "init" and "create" upcalls. By doing this, we don't
need to revise the upcall format. The UMH upcall can simply make use of
this info if it happens to be present. If it's not then it can just
avoid lifting the grace period early.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Upstream Id: d4318acd5d2d34d69a46537f057b20a8f0266e1e
Bugzilla: 1111712
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>