[fs] nfsd: serialize nfsdcltrack upcalls for a particular client
Message-id: <1423596680-6048-41-git-send-email-steved@redhat.com>
Patchwork-id: 105654
O-Subject: [RHEL7.2 PATCH 40/53] nfsd: serialize nfsdcltrack upcalls for a particular client
Bugzilla: 1111712
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

In a later patch, we want to add a flag that will allow us to reduce the
need for upcalls. In order to handle that correctly, we'll need to
ensure that racing upcalls for the same client can't occur. In practice
it should be rare for this to occur with a well-behaved client, but it
is possible.

Convert one of the bits in the cl_flags field to be an upcall bitlock,
and use it to ensure that upcalls for the same client are serialized.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Upstream Id: d682e750ce14cfb3be655e6d492c77511e637228
Bugzilla: 1111712
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>