[fs] gfs2: Allocate reservation during splice_write
Message-id: <1657315485.17668349.1424702502838.JavaMail.zimbra@redhat.com>
Patchwork-id: 106378
O-Subject: [RHEL7.2 PATCH] GFS2: Allocate reservation during splice_write
Bugzilla: 1193910
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch is for a bug that causes the kernel to crash whenever
sendfile() operations are done on GFS2. It was tested using a
reliable reproducer program in the bugzilla.
A patch was sent upstream to fix the problem, but due to changes
in the upstream kernel, the patch is slightly different.

Description:

This patch adds a GFS2-specific function for splice_write which
first calls function gfs2_rs_alloc to make sure a reservation
structure has been allocated before attempting to reserve blocks.

rhbz#1193910

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>