[x86] mm: Clean up the TLB flushing code
Message-id: <1424806302-20924-1-git-send-email-lwoodman@redhat.com>
Patchwork-id: 106416
O-Subject: [RHEL7.2 PATCH] Intel 7.2 FEATURE - Simplify TLB flushing code
Bugzilla: 1172891
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

The attached upstream backport simplifies the TLB flushing code.

We have two basic ways that the hardware lets us flush the TLB: one page at
a time, or flush the entire TLB. The kernel tries to choose the mode based
on the CPU's TLB size and micro-architecture from CPUID model/family.
However, investigation revealed that the data used for tuning
Ivybridge/Sandybridge were not reproducible. To fix this, we removed the
code which attempted to tune for the architecture and replaced it with a
much simpler tunable.

The result is greatly simplified code that does not have to be updated and
tested for every single family/model pair that Intel releases.

BZ: 1172891

Testing: Done by me and Intel.

Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=8775075

 commit 4995ab9cf512e9a6cc07dfd6b1d4e2fc48ce7fef
 Author: Dave Hansen <dave.hansen@linux.intel.com>
 Date:   Thu Jul 31 08:40:54 2014 -0700

    x86/mm: Clean up the TLB flushing code

    The

        if (cpumask_any_but(mm_cpumask(mm), smp_processor_id()) < nr_cpu_ids)

    line of code is not exactly the easiest to audit, especially when
    it ends up at two different indentation levels.  This eliminates
    one of the the copy-n-paste versions.  It also gives us a unified
    exit point for each path through this function.  We need this in
    a minute for our tracepoint.

    Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
    Link: http://lkml.kernel.org/r/20140731154054.44F1CDDC@viggo.jf.intel.com
    Acked-by: Rik van Riel <riel@redhat.com>
    Acked-by: Mel Gorman <mgorman@suse.de>
    Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>