[fs] pipe: fix pipe corruption and iovec overrun on partial copy
Message-id: <1425582109-11922-1-git-send-email-sjenning@redhat.com>
Patchwork-id: 107014
O-Subject: [RHEL7.2 PATCH] [fs] pipe: fix pipe corruption and iovec overrun on partial copy
Bugzilla: 1198843
CVE: CVE-2015-1805
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Lachlan McIlroy <lmcilroy@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1198843
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=8818413
RHEL6 BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1185166
Testing: Basic pipe stressing test in Beaker

When a pipe_read() or pipe_write() is called, it tries to copy to/from
the buffer using a kmap_atomic() mapping in an "atomic" mode.  However
if accessing the page that contains the iovec generates a page fault,
kmap_atomic() can't be used, and fallback mode with kmap() is used.

The problem is that pipe_iov_copy_from_user() and
pipe_iov_copy_to_user() can partially copy some of the iovec before
failing in the atomic mode and this progress is not tracked.  During the
retry, the entire copy is retried, but the iovecs that were processed in
the atomic mode have their iov->len set to 0 so they are not copied
again.  This causes a mismatch between the amount of data available in
the iovecs and the amount of data trying to be copied, resulting in an
overrun the iovec array, resulting in pipe and memory corruption.

This commit modifies the code so that the partial progress of the
attempted atomic mode copy is tracked so that the slow path retry has the
correct state.

Signed-off-by: Seth Jennings <sjenning@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>