[s390] zcrypt: enable s390 hwrng to seed kernel entropy
Message-id: <1426499979-9631-3-git-send-email-brueckner@redhat.com>
Patchwork-id: 107357
O-Subject: [RHEL7.2 PATCH 2/2] [s390] s390/zcrypt: enable s390 hwrng to seed kernel entropy
Bugzilla: 1196398
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Description
-----------
Author: Ingo Tuchscherer <ingo.tuchscherer@de.ibm.com>

    s390/zcrypt: enable s390 hwrng to seed kernel entropy

    Set the 'quality' property in the zcrypt rng device structure to enable the
    zcrypt hwrng device to take part in the kernel entropy seeding process.
    A module parameter named hwrng_seed will be introduced to disable the
    participation. By default this parameter is set to 1 (enabled).

    Signed-off-by: Ingo Tuchscherer <ingo.tuchscherer@de.ibm.com>

Bugzilla
--------
BZ 1196398
https://bugzilla.redhat.com/show_bug.cgi?id=1196398

Upstream status of the patch
----------------------------
This patch is not yet included in the upstream kernel.  But the change was
discussed as part of the common code hwrng entropy changes last year.  At that
point in time, the common code did not make it upstream either.  Now the
common hwrng entropy are upstream and backported to RHEL7 but the zcrypt part
is missing.  So here is the zcrytp part of it.

For details on the discussion, see https://lkml.org/lkml/2014/7/30/190

We plan to include this change in the next kernel release, v4.1.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=8847004

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>