[infiniband] ipoib: drop mcast_mutex usage
Message-id: <86aab14d3dc6bf29807eed8e76c967f8d0645dcf.1427392488.git.dledford@redhat.com>
Patchwork-id: 108741
O-Subject: [Patch RHEL7 10/10] IB/ipoib: drop mcast_mutex usage
Bugzilla: 1183881
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

We needed the mcast_mutex when we had to prevent the join completion
callback from having the value it stored in mcast->mc be overwritten
by a delayed return from ib_sa_join_multicast.  By storing the return
of ib_sa_join_multicast in an intermediate variable, we prevent a
delayed return from ib_sa_join_multicast from overwriting the valid
contents of mcast->mc, and we no longer need a mutex to force the
join callback to run after the return of ib_sa_join_multicast.  This
allows us to do away with the mutex entirely and protect our critical
sections with a just a spinlock instead.  This is highly desirable
as there were some places where we couldn't use a mutex because the
code was not allowed to sleep, and so we were currently using a mix
of mutex and spinlock to protect what we needed to protect.  Now we
only have a spin lock and the locking complexity is greatly reduced.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit d8b8d4bb38090319b57a9e2bb80dbfb7c77a8c50)
Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>