[vfio] vfio-pci: Attempt bus/slot reset on release
Message-id: <20150327212348.6042.53635.stgit@gimli.home>
Patchwork-id: 108933
O-Subject: [RHEL7.2 PATCH 10/24] vfio-pci: Attempt bus/slot reset on release
Bugzilla: 1206275
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1206275
Upstream: bc4fba77124e2fe4eb14bcb52875c0b0228deace

Each time a device is released, mark whether a local reset was
successful or whether a bus/slot reset is needed.  If a reset is
needed and all of the affected devices are bound to vfio-pci and
unused, allow the reset.  This is most useful when the userspace
driver is killed and releases all the devices in an unclean state,
such as when a QEMU VM quits.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>