[vfio] Add and use device request op for vfio bus drivers
Message-id: <20150327212429.6042.28859.stgit@gimli.home>
Patchwork-id: 108940
O-Subject: [RHEL7.2 PATCH 17/24] vfio: Add and use device request op for vfio bus drivers
Bugzilla: 1206275
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1206275
Upstream: 13060b64b819c194909121b90b5f8dd9abb5ea4e

When a request is made to unbind a device from a vfio bus driver,
we need to wait for the device to become unused, ie. for userspace
to release the device.  However, we have a long standing TODO in
the code to do something proactive to make that happen.  To enable
this, we add a request callback on the vfio bus driver struct,
which is intended to signal the user through the vfio device
interface to release the device.  Instead of passively waiting for
the device to become unused, we can now pester the user to give
it up.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>