[char] random: statically compute poolbitshift, poolbytes, poolbits
Message-id: <E1Yca5l-0001nB-4c@gondolin.me.apana.org.au>
Patchwork-id: 108952
O-Subject: [RHEL7 PATCH 1/3] random: statically compute poolbitshift, poolbytes, poolbits
Bugzilla: 1110044
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7 Bugzilla #1110044

Upstream commit: 9ed17b70b409dc48c134a80b5a6df582ba759de2

Use a macro to statically compute poolbitshift (will be used in a
subsequent patch), poolbytes, and poolbits.  On virtually all
architectures the cost of a memory load with an offset is the same as
the one of a memory load.

It is still possible for this to generate worse code since the C
compiler doesn't know the fixed relationship between these fields, but
that is somewhat unlikely.

Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Rafael Aquini <aquini@redhat.com>