[fs] dlm: adopt orphan locks
Message-id: <20150402211512.GA12530@redhat.com>
Patchwork-id: 109385
O-Subject: [RHEL7.2 PATCH] dlm: adopt orphan locks
Bugzilla: 1208288
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>

bug 1208288
build https://brewweb.devel.redhat.com/taskinfo?taskID=8940821
upstream http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/fs/dlm?id=2ab4bd8ea3a6954bc79a9bbeb291cd6c2d6213a7
tested myself

A process may exit, leaving an orphan lock in the lockspace.
This adds the capability for another process to acquire the
orphan lock.  Acquiring the orphan just moves the lock from
the orphan list onto the acquiring process's list of locks.

An adopting process must specify the resource name and mode
of the lock it wants to adopt.  If a matching lock is found,
the lock is moved to the caller's 's list of locks, and the
lkid of the lock is returned like the lkid of a new lock.

If an orphan with a different mode is found, then -EAGAIN is
returned.  If no orphan lock is found on the resource, then
-ENOENT is returned.  No async completion is used because
the result is immediately available.

Also, when orphans are purged, allow a zero nodeid to refer
to the local nodeid so the caller does not need to look up
the local nodeid.

Signed-off-by: David Teigland <teigland@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>