[drm] Missed clflushopt in drm_clflush_virt_range
Message-id: <20150407123912.4485.37779.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 109420
O-Subject: [PATCH RHEL7.2 BZ1170846 6/6] drm: Missed clflushopt in drm_clflush_virt_range
Bugzilla: 1170846
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1170846

Description:
------------
With this commit:

2a0788dc9bc4 x86: Use clflushopt in drm_clflush_virt_range

If clflushopt is available on the system, we use it instead of clflush
in drm_clflush_virt_range.  There were two calls to clflush in this
function, but only one was changed to clflushopt.  This patch changes
the other clflush call to clflushopt.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reported-by: Matthew Wilcox <matthew.r.wilcox@intel.com>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=7927096868061ff10fdb33b295dc0bcfebe2b35b

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>