[kernel] wait: Fix __wait_on_atomic_t() to call the action func if the counter != 0
Message-id: <20150407222251.21737.30955.stgit@warthog.procyon.org.uk>
Patchwork-id: 109453
O-Subject: [RHEL7 PATCH 01/50] Fix __wait_on_atomic_t() to call the action func if the counter != 0 [ver #2]
Bugzilla: 1130457 1204964
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Fix __wait_on_atomic_t() so that it calls the action func if the counter != 0
rather than if the counter is 0 so as to be analogous to __wait_on_bit().

Thanks to Yacine who found this by visual inspection.

This will affect FS-Cache in that it will could fail to sleep correctly when
trying to clean up after a netfs cookie is withdrawn.

Reported-by: Yacine Belkadi <yacine.belkadi.1@gmail.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
cc: Milosz Tanski <milosz@adfin.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Original-commit: 42577ca8c3616baaafdd8f167b2e1fb959026081
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>