[fs] fscache: remove spin_lock() from the condition in while()
Message-id: <20150407222301.21737.57632.stgit@warthog.procyon.org.uk>
Patchwork-id: 109455
O-Subject: [RHEL7 PATCH 03/50] fs/fscache: remove spin_lock() from the condition in while() [ver #2]
Bugzilla: 1130457 1204964
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Sebastian Andrzej Siewior <bigeasy@linutronix.de>

The spinlock() within the condition in while() will cause a compile error
if it is not a function. This is not a problem on mainline but it does not
look pretty and there is no reason to do it that way.
That patch writes it a little differently and avoids the double condition.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: David Howells <dhowells@redhat.com>
Tested-By: Milosz Tanski <milosz@adfin.com>
Acked-by: Jeff Layton <jlayton@redhat.com>
Original-commit: ee8be57bc331f00d520c4c8a78ffa9590ed41c2c
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>