[fs] cachefiles: Fix memory leak in cachefiles_check_auxdata error paths
Message-id: <20150407222423.21737.39633.stgit@warthog.procyon.org.uk>
Patchwork-id: 109469
O-Subject: [RHEL7 PATCH 17/50] CacheFiles: Fix memory leak in cachefiles_check_auxdata error paths [ver #2]
Bugzilla: 1130457 1204964
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Josh Boyer <jwboyer@redhat.com>

In cachefiles_check_auxdata(), we allocate auxbuf but fail to free it if
we determine there's an error or that the data is stale.

Further, assigning the output of vfs_getxattr() to auxbuf->len gives
problems with checking for errors as auxbuf->len is a u16.  We don't
actually need to set auxbuf->len, so keep the length in a variable for
now.  We shouldn't need to check the upper limit of the buffer as an
overflow there should be indicated by -ERANGE.

While we're at it, fscache_check_aux() returns an enum value, not an
int, so assign it to an appropriately typed variable rather than to ret.

Signed-off-by: Josh Boyer <jwboyer@fedoraproject.org>
Signed-off-by: David Howells <dhowells@redhat.com>
cc: Hongyi Jia <jiayisuse@gmail.com>
cc: Milosz Tanski <milosz@adfin.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Original-commit: 607566aeccde6ffccde5eef173ed0d277eec4a2d
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>