[fs] fs-cache: Handle removal of unadded object to the fscache_object_list rb tree
Message-id: <20150407222452.21737.35002.stgit@warthog.procyon.org.uk>
Patchwork-id: 109474
O-Subject: [RHEL7 PATCH 22/50] FS-Cache: Handle removal of unadded object to the fscache_object_list rb tree [ver #2]
Bugzilla: 1130457 1204964
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

When FS-Cache allocates an object, the following sequence of events can
occur:

 -->fscache_alloc_object()
    -->cachefiles_alloc_object() [via cache->ops->alloc_object]
    <--[returns new object]
    -->fscache_attach_object()
    <--[failed]
    -->cachefiles_put_object() [via cache->ops->put_object]
       -->fscache_object_destroy()
          -->fscache_objlist_remove()
             -->rb_erase() to remove the object from fscache_object_list.

resulting in a crash in the rbtree code.

The problem is that the object is only added to fscache_object_list on
the success path of fscache_attach_object() where it calls
fscache_objlist_add().

So if fscache_attach_object() fails, the object won't have been added to
the objlist rbtree.  We do, however, unconditionally try to remove the
object from the tree.

Thanks to NeilBrown for finding this and suggesting this solution.

Reported-by: NeilBrown <neilb@suse.de>
Signed-off-by: David Howells <dhowells@redhat.com>
Tested-by: (a customer of) NeilBrown <neilb@suse.de>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Original-commit: 7026f1929e18921fd67bf478f475a8fdfdff16ae
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>